+++ examples/test.py
@@ -1,9 +1,12 @@
 #!/usr/bin/env python3
-"""Cloud example.
+"""Local Example.
 
 How to use this script:
-    export CLIENT_ID="dnfqsdfjqsjfqsdjfqf"
-    export CLIENT_SECRET="djfqsdkfjqsdkfjqsdkfjqsdkfjkqsdjfkjdkfqjdskf"
+    pip install python-dotenv
+    export LOCALGW='192.168.0.2'
+    export USER_NAME="dnfqsdfjqsjfqsdjfqf"
+    export PASSWORD="djfqsdkfjqsdkfjqsdkfjqsdkfjkqsdjfkjdkfqjdskf"
+    export TLS=True
     python cloud_example.py
 """
 from __future__ import annotations
@@ -18,15 +21,7 @@ except ModuleNotFoundError as exc:
     msg = "You have to run 'pip install python-dotenv' first"
     raise ImportError(msg) from exc
 
-try:
-    from authlib.integrations.httpx_client import AsyncOAuth2Client
-except ModuleNotFoundError as exc:
-    msg = "You have to run 'pip install httpx authlib' first"
-    raise ImportError(msg) from exc
-
-
-from pyhaopenmotics import OpenMoticsCloud
-from pyhaopenmotics.const import CLOUD_SCOPE, OAUTH2_TOKEN
+from pyhaopenmotics import LocalGateway
 
 # UNCOMMENT THIS TO SEE ALL THE HTTPX INTERNAL LOGGING
 log = logging.getLogger()
@@ -40,38 +35,51 @@ log.addHandler(console)
 
 load_dotenv()
 
-client_id = os.environ["CLIENT_ID"]
-client_secret = os.environ["CLIENT_SECRET"]
+localgw = os.environ["LOCALGW"]
+username = os.environ["USER_NAME"]
+password = os.environ["PASSWORD"]
+port = int(os.environ["PORT"])
 
 
 async def main() -> None:
-    """Docstring."""
-    token = None
+    """Show example on controlling your OpenMotics device."""
+    async with LocalGateway(
+        localgw=localgw,
+        username=username,
+        password=password,
+        port=port,
+    ) as omclient:
+        await omclient.exec_action("get_version")
 
-    async with AsyncOAuth2Client(
-        client_id=client_id,
-        client_secret=client_secret,
-        token_endpoint_auth_method="client_secret_post",  # noqa: S106 # nosec
-        scope=CLOUD_SCOPE,
-        token_endpoint=OAUTH2_TOKEN,
-        grant_type="client_credentials",
-    ) as httpx_session:
-        token = await httpx_session.fetch_token(
-            url=OAUTH2_TOKEN,
-            grant_type="client_credentials",
-        )
-        access_token = token.get("access_token")
+        outputs = await omclient.outputs.get_all()
 
-        omclient = OpenMoticsCloud(token=access_token)
+        # if outputs[0].status.on is True:
+        #     pass
+        # else:
+        #     pass
 
-        installations = await omclient.installations.get_all()
+        # await omclient.outputs.get_by_id(0)
 
-        i_id = installations[0].idx
+        #         await omclient.inputs.get_all()
 
-        await omclient.installations.get_by_id(i_id)
-        omclient.installation_id = i_id
+        #        await omclient.sensors.get_all()
 
-        await omclient.outputs.get_by_id(99351)
+        #        await omclient.shutters.get_all()
+
+        #        await omclient.groupactions.get_all()
+
+        #        await omclient.thermostats.groups.get_all()
+
+        #        await omclient.thermostats.units.get_all()
+
+        print("New API -----------------------------------------------------------")
+
+        # await omclient.get('/api/thermostat/control')
+        await omclient.get("/api/thermostat/config")
+
+        await omclient.get("/api/thermostat/config/2")
+
+        await omclient.get("/api/sensors")
 
         await omclient.close()
