+++ src/pyhaopenmotics/__init__.py
@@ -1,17 +1,16 @@
 """Module HTTP communication with the OpenMotics API."""
 
-from pyhaopenmotics.cloud.models import Installation
-
-from .errors import (
+from pyhaopenmotics.client.errors import (
     AuthenticationError,
     OpenMoticsConnectionError,
     OpenMoticsConnectionSslError,
     OpenMoticsConnectionTimeoutError,
     OpenMoticsError,
 )
-from .helpers import get_ssl_context
-from .localgateway import LocalGateway
-from .openmoticscloud import OpenMoticsCloud
+from pyhaopenmotics.client.localgateway import LocalGateway
+from pyhaopenmotics.client.openmoticscloud import OpenMoticsCloud
+from pyhaopenmotics.client.websocket import WebsocketClient
+from pyhaopenmotics.cloud.models import Installation
 
 __all__ = [
     "OpenMoticsCloud",
@@ -22,5 +21,6 @@ __all__ = [
     "OpenMoticsConnectionSslError",
     "AuthenticationError",
     "Installation",
-    "get_ssl_context",
+    "WebsocketClient",
+    # "get_ssl_context",
 ]
