+++ src/pyhaopenmotics/__version__.py
@@ -1,3 +1,3 @@
 """Asynchronous Python client the OpenMotics API."""
 
-__version__ = "0.0.5"
+__version__ = "0.0.6"
