+++ src/pyhaopenmotics/client/errors.py
